+++ src/main/java/com/careerup/careerupspring/dto/UserDTO.java
@@ -1,8 +1,6 @@
 package com.careerup.careerupspring.dto;
 
 import com.careerup.careerupspring.entity.UserEntity;
-import com.careerup.careerupspring.entity.UserFieldEntity;
-import com.careerup.careerupspring.entity.UserSkillEntity;
 import lombok.*;
 
 import java.util.List;
@@ -18,11 +16,16 @@ public class UserDTO {
     private String profile;
     private String nickname;
     private String company;
+    private UserEntity.roleType roleType;
     private List<String> fields;
     private List<String> skills;
 
     public UserEntity toEntity(){
         UserEntity build = UserEntity.builder()
+                .email(email)
+                .roleType(roleType)
+                .password(password)
+                .company(company)
                 .id(id).profile(profile).build();
         return build;
     }
