+++ src/main/java/com/careerup/careerupspring/repository/UserRepository.java
@@ -10,6 +10,7 @@ import java.util.UUID;
 
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, UUID>, UserRepositoryCustom {
+    public boolean existsByNickname(String nickname);
     public Optional<UserEntity> findByEmail(String email);
     public Optional<UserEntity> findByNickname(String nickname);
 }
